% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata_lasso_binomial}
\alias{simdata_lasso_binomial}
\title{Simulated complex survey data}
\format{
\subsection{\code{simdata_lasso_binomial}}{

A data frame with 1,720 rows and 54 columns:
\describe{
\item{y}{dichotomous response variable}
\item{x.1,...,x.50}{covariates}
\item{strata}{strata identifiers}
\item{cluster}{cluster identifiers}
\item{weights}{sampling weights}
}
}
}
\usage{
simdata_lasso_binomial
}
\description{
Simulated complex survey data
}
\keyword{datasets}
