% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wse.R
\name{wse}
\alias{wse}
\title{Estimation of the sensitivity with complex survey data}
\usage{
wse(
  response.var,
  phat.var,
  weights.var = NULL,
  tag.event = NULL,
  cutoff.value,
  data = NULL,
  design = NULL
)
}
\arguments{
\item{response.var}{Vector with information of the response variable
or character string with the name of the variable
indicating the response variable in the data set.}

\item{phat.var}{Vector with information of the estimated probabilities
or character string with the name of the variable
indicating these probabilities in the data set.}

\item{weights.var}{Vector with information of the sampling weights
or character string with the name of the variable
indicating the sampling weights in the data set.
If \code{NULL}, the sampling design should be specified in the argument \code{design}.
For unweighted estimates, set all the sampling weight values to 1.}

\item{tag.event}{Label indicating the event of interest in \code{response.var}.
The default option is \code{tag.event = NULL}, which selects the class with the lowest number of units.}

\item{cutoff.value}{A numeric value indicating the cut-off point to be used.
No default value for this argument. A numeric value must be indicated necessarily.}

\item{data}{Data frame which must incorporate information on the columns
\code{response.var}, \code{phat.var} and \code{weights.var}.
If \code{data=NULL}, then specific numerical vectors must be included in
\code{response.var}, \code{phat.var} and \code{weights.var}
or the sampling design should be indicated in the argument \code{design}.}

\item{design}{An object of class \code{survey.design} generated by
\code{survey::svydesign} indicating the complex sampling design of the data.
If \code{NULL} information on the data and weights must be included in the argument
\code{data} or as a vector in the argument \code{weights.var}.}
}
\value{
The output of this function is a list of 4 elements containing the following information:
\itemize{
\item \code{Sew}: a numeric value indicating the weighted estimate of the sensitivity parameter.
\item \code{tags}: list containing one element with the following information:
\itemize{
\item \code{tag.event}: a character string indicating the label used to indicate event of interest.
}
\item \code{basics}: a list containing information of the following 6 elements:
\itemize{
\item \code{n}: a numeric value indicating the number of units in the data set.
\item \code{n.event}: a numeric value indicating the number of units in the data set with the event of interest.
\item \code{n.event.class}: a numeric value indicating the number of units in the data set with the event of interest that are correctly classified as events based on the selected cut-off point.
\item \code{hatN}: number of units in the population, represented by all the units in the data set, i.e., the sum of the sampling weights of the units in the data set.
\item \code{hatN.event}: number of units with the event of interest represented in the population by all the event units in the data set, i.e., the sum of the sampling weights of the units with the event of interest in the data set.
\item \code{hatN.event.class}: number of event units represented in the population by the event units in the data set that have been correctly classified as events based on the selected cut-off point, i.e., the sum of the sampling weights of the correctly classified event units in the data set.
}
\item \code{call}: an object saving the information about the way in which the function has been run.
}
}
\description{
Estimate the sensitivity parameter for a given cut-off point
considering sampling weights with complex survey data.
}
\examples{
data(example_data_wroc)

se.obj <- wse(response.var = "y", phat.var = "phat", weights.var = "weights",
              tag.event = 1, cutoff.value = 0.5, data = example_data_wroc)

# Or equivalently
se.obj <- wse(response.var = example_data_wroc$y,
              phat.var = example_data_wroc$phat,
              weights.var = example_data_wroc$weights,
              tag.event = 1, cutoff.value = 0.5)
}
